requirements.txt
dotenv>=0.9.9
numpy>=2.2.3
scipy>=1.15.2
shapely>=2.0.7
lark>=1.2.2
lxml>=5.3.1
discord>=2.3.2
black>=25.1.0
pillow>=12.0.0
scikit-image>=0.25.2